moved main part of ingest to core
